Avoid creating threads for small jobs

+++ src/utils/index.js
@@ -1,6 +1,7 @@
 import { quantileSorted } from 'd3-array'
 import { color } from 'd3-color'
 import { scaleLinear } from 'd3-scale'
+import { Parser } from 'expr-eval'
 import PropTypes from 'prop-types'
 import * as R from 'ramda'
 import { GenIcon } from 'react-icons'
@@ -238,6 +239,51 @@ const promiseAllObject = (obj) =>
   Promise.all(R.values(obj)).then(R.zipObj(R.keys(obj)))
 
 export const calculateStatAnyDepth = (valueBuffers) => {
+  const valueLists = R.map((buffer) => new Float64Array(buffer))(valueBuffers)
+  const parser = new Parser()
+  const calculate = (group, calculation) => {
+    // if there are no calculations just return values at the group indicies
+    if (R.has(calculation, valueLists)) {
+      return R.pipe((d) => d[calculation], R.pick(group), R.values)(valueLists)
+    }
+    // define groupSum for each base level group
+    const preSummed = {}
+    parser.functions.groupSum = (statName) => {
+      // groupSum only works for non-derived stats
+      // dont recalculate sum for each stat
+      if (R.isNil(preSummed[statName])) {
+        preSummed[statName] = R.sum(
+          R.map((idx) => valueLists[statName][idx], group)
+        )
+      }
+      return preSummed[statName]
+    }
+    return group.map((idx) => {
+      const proxy = new Proxy(valueLists, {
+        get(target, name, receiver) {
+          return Reflect.get(target, name, receiver)[idx]
+        },
+      })
+      try {
+        return parser.parse(calculation).evaluate(
+          // evaluate each list item
+          proxy
+        )
+      } catch {
+        console.warn(`Malformed calculation: ${calculation}`)
+        // if calculation is malformed return simplified array
+        return parseArray(
+          parser
+            .parse(calculation)
+            .simplify(
+              // evaluate each list item
+              proxy
+            )
+            .toString()
+        )
+      }
+    })
+  }
   const group = async (groupBys, calculation, indicies) => {
     const currentGroupBy = groupBys[0]
     const keyFn = R.pipe(currentGroupBy, R.join(' \u279D '))
@@ -258,23 +304,25 @@ export const calculateStatAnyDepth = (valueBuffers) => {
         ? async (d) =>
             await promiseAllObject(
               R.map((group) => {
-                return new Promise((resolve, reject) => {
-                  const worker = new Worker(
-                    new URL('./computationWorker.js', import.meta.url)
-                  )
-                  worker.postMessage({
-                    indicies: group,
-                    calculation,
-                    valueBuffers,
+                if (group.length < 1000) return calculate(group, calculation)
+                else
+                  return new Promise((resolve, reject) => {
+                    const worker = new Worker(
+                      new URL('./computationWorker.js', import.meta.url)
+                    )
+                    worker.postMessage({
+                      indicies: group,
+                      calculation,
+                      valueBuffers,
+                    })
+                    worker.onmessage = (e) => {
+                      worker.terminate()
+                      resolve(e.data)
+                    }
+                    worker.onerror = (e) => {
+                      reject(e)
+                    }
                   })
-                  worker.onmessage = (e) => {
-                    worker.terminate()
-                    resolve(e.data)
-                  }
-                  worker.onerror = (e) => {
-                    reject(e)
-                  }
-                })
               })(d)
             )
         : async (d) =>
